+++ pca9685.h
@@ -10,5 +10,5 @@
 void init_pca9685(i2c_device* dev);
 void setfreq_pca9685(i2c_device* dev, unsigned int freq);
 
-void setpwm_pca9685(i2c_device* dev, uint8_t pin, uint8_t on_percent, uint8_t delay_percent);
+void setpwm_pca9685(i2c_device* dev, uint8_t pin, float on_percent, uint8_t delay_percent);
